pyproject.toml
[project]
name = "landing-genie"
version = "0.1.0"
description = "AI assisted landing page generator using Gemini CLI and Cloudflare Pages"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]",
    "python-dotenv",
    "requests"
]

[project.scripts]
landing-genie = "landing_genie.cli:app"

[tool.setuptools.packages.find]
include = ["landing_genie*"]
exclude = ["prompts", "sites"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
